Move all libraries into lib
Everything except for //software/devices and GTest code has been
moved under //lib.
