+++ lib/arduino/FastLedManager.hpp
@@ -4,8 +4,8 @@
 #include <Types.hpp>
 #include <vector>
 
-#include "../../lib/effect/Effect.hpp"
-#include "../../software/generic/DeviceDescription.hpp"
+#include "../effect/Effect.hpp"
+#include "../device/DeviceDescription.hpp"
 #include "../generic/LedManager.hpp"
 
 class FastLedManager : public LedManager {
