+++ lib/arduino/RadioHeadRadio.hpp
@@ -3,7 +3,7 @@
 
 #include <RH_RF69.h>
 
-#include "../generic/Radio.hpp"
+#include "../radio/Radio.hpp"
 
 const int kMaxFifoSizePacketSize = 64;
 const int kRadioOverhead = 3;
