+++ lib/color/ColorPalette.hpp
