+++ lib/device/DeviceDescription.cpp
