+++ lib/device/DeviceDescription.hpp
