+++ lib/effect/SwingingLights.cpp
@@ -1,7 +1,7 @@
 #include "SwingingLights.hpp"
 
-#include "../../software/generic/ColorPalette.hpp"
-#include "../../software/generic/Math.hpp"
+#include "../color/ColorPalette.hpp"
+#include "../math/Math.hpp"
 
 static const uint16_t SPREAD = MAX_UINT16 * 0.2;
