+++ lib/network/NetworkManager.cpp
