+++ platformio.ini
@@ -17,7 +17,7 @@ platform = https://github.com/candykingdom/platform-candykingdomsam.git
 board = rfboard
 framework = arduino
 monitor_speed = 115200
-build_src_filter = +<*> -<*/test/*> -<*/build/*> -<devices> +<devices/node>
+build_src_filter = +<lib> +<software/devices/node>
 platform_packages = 
 ; Choose the correct package for your OS:
 	tool-bossac-candykingdom @ https://candykingdom.github.io/firefly-v2-board/tools/bossac_linux.tar.bz2
@@ -32,7 +32,7 @@ platform = espressif32
 board = esp32thing_plus
 framework = arduino
 monitor_speed = 115200
-build_src_filter = +<*> -<*/test/*> -<devices> +<devices/dmx>
+build_src_filter = +<lib> +<software/devices/dmx>
 build_flags = 
 	-D RADIO_SS=32
 	-D RADIO_DIO=14
