+++ software/devices/range_test/range_test.ino
@@ -1,9 +1,9 @@
 #undef max
 #undef min
-#include "../../arduino/FastLedManager.hpp"
-#include "../../arduino/RadioHeadRadio.hpp"
-#include "../../generic/NetworkManager.hpp"
-#include "../../generic/RadioStateMachine.hpp"
+#include "../../../lib/arduino/FastLedManager.hpp"
+#include "../../../lib/arduino/RadioHeadRadio.hpp"
+#include "../../../lib/network/NetworkManager.hpp"
+#include "../../../lib/radio/RadioStateMachine.hpp"
 
 const int kNumLeds = 60;
