+++ software/generic/test/FakeNetwork.cpp
@@ -3,8 +3,8 @@
 #include <Debug.hpp>
 #include <cstdio>
 
-#include "../DeviceDescription.hpp"
-#include "../NetworkManager.hpp"
+#include "../../lib/device/DeviceDescription.hpp"
+#include "../../lib/network/NetworkManager.hpp"
 
 //#define DEBUG
