+++ software/generic/test/FakeNetwork.hpp
@@ -1,8 +1,8 @@
 #ifndef __FAKE_NETWORK_H__
 #define __FAKE_NETWORK_H__
 
-#include "../Radio.hpp"
-#include "../RadioStateMachine.hpp"
+#include "../../lib/radio/Radio.hpp"
+#include "../../lib/radio/RadioStateMachine.hpp"
 #include "FakeLedManager.hpp"
 #include "FakeRadio.hpp"
