+++ software/generic/test/NetworkManagerTest.cpp
@@ -1,5 +1,5 @@
-#include "../NetworkManager.hpp"
-#include "../Radio.hpp"
+#include "../../lib/network/NetworkManager.hpp"
+#include "../../lib/radio/Radio.hpp"
 #include "FakeRadio.hpp"
 #include "gtest/gtest.h"
