+++ software/generic/test/RadioStateIntegrationTest.cpp
@@ -1,6 +1,6 @@
-#include "../NetworkManager.hpp"
-#include "../Radio.hpp"
-#include "../RadioStateMachine.hpp"
+#include "../../lib/network/NetworkManager.hpp"
+#include "../../lib/radio/Radio.hpp"
+#include "../../lib/radio/RadioStateMachine.hpp"
 #include "FakeNetwork.hpp"
 #include "FakeRadio.hpp"
 #include "RadioIntegrationTest.hpp"
